bonus(register): refactor public routes
